@@ -6,7 +6,7 @@ publish_to: "none"
 version: 0.0.1+1
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: "3.1.2"
 
 dependencies:
   
