@@ -6,7 +6,7 @@ const kUserMock = User(
   avatarUrl:
       'https://media.licdn.com/dms/image/C4D0BAQFGUHRJ26bFDw/company-logo_200_200/0/1519856619408?e=1699488000&v=beta&t=ZmHH9S3D3olLSwAS_Egw862invpc80wKT_Gd0C3n2A0',
   htmlUrl: 'htmlUrl',
-  name: 'FIAP',
+  name: 'Davi',
   company: 'FIAP',
   blog: 'https://www.linkedin.com/school/fiap/mycompany/verification/',
   location: 'São Paulo, SP',
