@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:github_discover/src/constants/mock/repository_mock.dart';
 import 'package:github_discover/src/constants/spacings.dart';
 import 'package:github_discover/src/domain/entities/repository.dart';
 import 'package:github_discover/src/presentation/components/app_bar.dart';
@@ -7,17 +6,25 @@ import 'package:github_discover/src/presentation/pages/repositories/widgets/repo
 import 'package:github_discover/src/utils/extensions/build_context_extensions.dart';
 import 'package:github_discover/src/utils/extensions/theme_data_extensions.dart';
 
-class RepositoryDetailsPage extends StatelessWidget {
-  const RepositoryDetailsPage({super.key});
+class RepositoryDetailsPage extends StatefulWidget {
+  final Repository repository;
+
+  const RepositoryDetailsPage({
+    super.key,
+    required this.repository,
+  });
 
   @override
-  Widget build(BuildContext context) {
-    Repository repository = kRepositoryMock;
+  State<RepositoryDetailsPage> createState() => _RepositoryDetailsPageState();
+}
 
+class _RepositoryDetailsPageState extends State<RepositoryDetailsPage> {
+  @override
+  Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: context.colors.kBackgrounDefaultColor,
       appBar: CustomAppBar(
-        text: repository.name?.toUpperCase(),
+        text: widget.repository.name?.toUpperCase(),
         hasBackButton: true,
       ),
       extendBodyBehindAppBar: true,
@@ -25,7 +32,7 @@ class RepositoryDetailsPage extends StatelessWidget {
         child: Container(
           padding: const EdgeInsets.all(Spacing.s24),
           width: MediaQuery.sizeOf(context).width,
-          child: RepositoryCard(repository: repository),
+          child: RepositoryCard(repository: widget.repository),
         ),
       ),
     );
