@@ -1,20 +1,35 @@
+// ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:flutter/material.dart';
-import 'package:github_discover/src/config/routes.dart';
-import 'package:github_discover/src/constants/mock/repository_mock.dart';
 import 'package:github_discover/src/constants/spacings.dart';
 import 'package:github_discover/src/constants/theme.dart';
 import 'package:github_discover/src/constants/typographies.dart';
+import 'package:github_discover/src/domain/entities/repository.dart';
 import 'package:github_discover/src/presentation/components/app_bar.dart';
 import 'package:github_discover/src/presentation/components/app_bar_bottom.dart';
 import 'package:github_discover/src/presentation/components/text.dart';
 import 'package:github_discover/src/presentation/pages/repositories/widgets/repository_list_tile.dart';
 import 'package:github_discover/src/utils/extensions/build_context_extensions.dart';
 import 'package:github_discover/src/utils/extensions/theme_data_extensions.dart';
-import 'package:go_router/go_router.dart';
 import 'package:intl/intl.dart';
 
-class RepositoriesSearchPage extends StatelessWidget {
-  const RepositoriesSearchPage({super.key});
+class RepositoriesSearchPage extends StatefulWidget {
+  final List<Repository> repository;
+  final Function(String) onSearch;
+  final void Function(Repository) onDetailTap;
+
+  const RepositoriesSearchPage({
+    super.key,
+    required this.repository,
+    required this.onSearch,
+    required this.onDetailTap,
+  });
+
+  @override
+  State<RepositoriesSearchPage> createState() => _RepositoriesSearchPageState();
+}
+
+class _RepositoriesSearchPageState extends State<RepositoriesSearchPage> {
+  String _searchString = "";
 
   @override
   Widget build(BuildContext context) {
@@ -24,8 +39,12 @@ class RepositoriesSearchPage extends StatelessWidget {
       backgroundColor: context.colors.kBackgrounDefaultColor,
       appBar: CustomAppBar(
         bottom: CustomAppBarBottom(
-          onChanged: (search) {},
-          onPressed: () {},
+          onChanged: (search) {
+            setState(() {
+              _searchString = search;
+            });
+          },
+          onPressed: () => widget.onSearch(_searchString),
         ),
       ),
       extendBodyBehindAppBar: true,
@@ -40,7 +59,8 @@ class RepositoriesSearchPage extends StatelessWidget {
               mainAxisSize: MainAxisSize.min,
               children: [
                 CustomText(
-                  text: context.locales.resultHeader(formatter.format(3)),
+                  text: context.locales
+                      .resultHeader(formatter.format(widget.repository.length)),
                   textAlign: TextAlign.start,
                   style: TypographyType.header,
                 ),
@@ -48,12 +68,13 @@ class RepositoriesSearchPage extends StatelessWidget {
                 ListView.builder(
                   physics: const ClampingScrollPhysics(),
                   shrinkWrap: true,
-                  itemCount: 3,
+                  itemCount: widget.repository.length,
                   itemBuilder: (context, index) {
+                    var repostory = widget.repository[index];
                     return RepositoryListTile(
-                      repository: kRepositoryMock,
+                      repository: repostory,
                       onTap: () {
-                        context.goNamed(AppRoute.repositoryDetails.name);
+                        widget.onDetailTap(widget.repository[index]);
                       },
                     );
                   },
