@@ -1,26 +1,34 @@
 import 'package:flutter/material.dart';
-import 'package:github_discover/src/config/routes.dart';
-import 'package:github_discover/src/constants/mock/user_mock.dart';
 import 'package:github_discover/src/constants/spacings.dart';
 import 'package:github_discover/src/constants/theme.dart';
 import 'package:github_discover/src/constants/typographies.dart';
+import 'package:github_discover/src/domain/entities/user.dart';
 import 'package:github_discover/src/presentation/components/app_bar.dart';
 import 'package:github_discover/src/presentation/components/app_bar_bottom.dart';
 import 'package:github_discover/src/presentation/components/text.dart';
 import 'package:github_discover/src/presentation/pages/users/widgets/user_list_tile.dart';
 import 'package:github_discover/src/utils/extensions/build_context_extensions.dart';
 import 'package:github_discover/src/utils/extensions/theme_data_extensions.dart';
-import 'package:go_router/go_router.dart';
 import 'package:intl/intl.dart';
 
 class UsersSearchPage extends StatefulWidget {
-  const UsersSearchPage({super.key});
+  final List<User> users;
+  final void Function(String) onSearch;
+  final void Function(User) onDetailTap;
+
+  const UsersSearchPage({
+    super.key,
+    required this.users,
+    required this.onSearch,
+    required this.onDetailTap,
+  });
 
   @override
   State<UsersSearchPage> createState() => _UsersSearchPageState();
 }
 
 class _UsersSearchPageState extends State<UsersSearchPage> {
+  String _searchString = "";
   var formatter = NumberFormat.compact(locale: locale);
 
   @override
@@ -29,8 +37,12 @@ class _UsersSearchPageState extends State<UsersSearchPage> {
       backgroundColor: context.colors.kBackgrounDefaultColor,
       appBar: CustomAppBar(
         bottom: CustomAppBarBottom(
-          onChanged: (search) {},
-          onPressed: () {},
+          onChanged: (search) {
+            setState(() {
+              _searchString = search;
+            });
+          },
+          onPressed: () => widget.onSearch(_searchString),
         ),
       ),
       extendBodyBehindAppBar: true,
@@ -45,7 +57,8 @@ class _UsersSearchPageState extends State<UsersSearchPage> {
               mainAxisSize: MainAxisSize.min,
               children: [
                 CustomText(
-                  text: context.locales.resultHeader(formatter.format(3)),
+                  text: context.locales
+                      .resultHeader(formatter.format(widget.users.length)),
                   textAlign: TextAlign.start,
                   style: TypographyType.header,
                 ),
@@ -63,15 +76,13 @@ class _UsersSearchPageState extends State<UsersSearchPage> {
                     physics: const ClampingScrollPhysics(),
                     scrollDirection: Axis.horizontal,
                     shrinkWrap: true,
-                    itemCount: 20,
+                    itemCount: widget.users.length,
                     itemBuilder: (context, index) {
                       return Flexible(
                         child: UserListTile(
-                          user: kUserMock,
-                          onTap: () {
-                            context.goNamed(AppRoute.userDetails.name);
-                          },
-                        ),
+                            user: widget.users[index],
+                            onTap: () =>
+                                widget.onDetailTap(widget.users[index])),
                       );
                     },
                   ),
